move ExecutionResultDTO to dto package
