+++ domain/dto/__init__.py
@@ -1,7 +1,14 @@
 from .company_dto import CompanyDTO
+from .execution_result_dto import ExecutionResultDTO
+from .metrics_dto import MetricsDTO
 from .nsd_dto import NSDDTO
 from .page_result_dto import PageResultDTO
-from .raw_company_dto import CodeDTO, CompanyDetailDTO, CompanyListingDTO, CompanyRawDTO
+from .raw_company_dto import (
+    CodeDTO,
+    CompanyDetailDTO,
+    CompanyListingDTO,
+    CompanyRawDTO,
+)
 from .sync_companies_result_dto import SyncCompaniesResultDTO
 
 __all__ = [
@@ -11,6 +18,8 @@ __all__ = [
     "CompanyListingDTO",
     "CompanyDetailDTO",
     "CodeDTO",
+    "ExecutionResultDTO",
+    "MetricsDTO",
     "PageResultDTO",
     "SyncCompaniesResultDTO",
 ]
