+++ domain/ports/__init__.py
@@ -1,14 +1,11 @@
+from domain.dto import ExecutionResultDTO, MetricsDTO
+
 from .company_repository_port import CompanyRepositoryPort
 from .company_source_port import CompanySourcePort
 from .metrics_collector_port import MetricsCollectorPort
 from .nsd_repository_port import NSDRepositoryPort
 from .nsd_source_port import NSDSourcePort
-from .worker_pool_port import (
-    ExecutionResultDTO,
-    LoggerPort,
-    MetricsDTO,
-    WorkerPoolPort,
-)
+from .worker_pool_port import LoggerPort, WorkerPoolPort
 
 __all__ = [
     "WorkerPoolPort",
